+++ test/int/project_test.go
@@ -1,4 +1,4 @@
-package project
+package int
 
 import (
 	"context"
